Fix new players loop to only record the playername
The loop was serializing all the form elements in Safari, including the submit button name, so we were getting two players in the system, one with the player's actual name, and another just called "Add Player" (the text of the submit button. We just put an if check to make sure it's only doing "playername" field. Huzzah.