Switch UI to dollar-based
Previously you choose a name and a dollar value to track things, but when tested in actual game-play, it is too hard to remember what the dollar value was after answering the question. The new UI shows the dollar values and allows you to assign a name, so when playing, before they announce the question you can choose the dollar value and then, if someone answers it correctly, you just choose them from the list. It's much easier! The code is BS but it works ...